requirements.txt
Flask
Werkzeug
mysql_connector_python
requests
sortedcontainers
urllib3
